Moved game objects to their own directory
