+++ game/objects/base.py
