+++ game/objects/sector.py
